Rename shared/content.py to shared/message.py
