+++ pybadge/discord_display.py
@@ -2,7 +2,7 @@ import displayio
 from adafruit_display_text.label import Label
 from adafruit_bitmap_font import bitmap_font
 import shared.layout_helper as layout
-import shared.content as content
+import shared.messages as messages
 
 # Load the title font
 TITLE_FONTNAME = "/fonts/cherry-13-b.bdf"
@@ -12,7 +12,7 @@ TITLE_FONT = bitmap_font.load_font(TITLE_FONTNAME)
 MESSAGE_FONTNAME = "/fonts/cherry-11-r.bdf"
 MESSAGE_FONT = bitmap_font.load_font(MESSAGE_FONTNAME)
 
-class DiscordMessageGroup(displayio.Group, content.DiscordMessageBase):
+class DiscordMessageGroup(displayio.Group, messages.DiscordMessageBase):
     """Display class for Discord messages"""
 
     max_lines = 5
@@ -20,7 +20,7 @@ class DiscordMessageGroup(displayio.Group, content.DiscordMessageBase):
     def __init__(self, message: str, user: str, cmd_type: int, dark_mode: bool = True) -> None:
 
         super().__init__()
-        content.DiscordMessageBase.__init__(self, message, user, cmd_type)
+        messages.DiscordMessageBase.__init__(self, message, user, cmd_type)
 
         text_color = 0xFFFFFF if dark_mode else 0x000000
         self.username_label = Label(TITLE_FONT, text=user, color=text_color, y=8)
